[#50] Ref: 로그인시 uid를 저장하는 UserDefaults를 파일로 분리
